Add hunger, food, improve item system, fix farming block ids
